Added contracts entity

+++ app/src/main/java/t6proj/employees/persistence/dao/EmployeeDao.java
@@ -41,7 +41,7 @@ public class EmployeeDao {
         return new PaginatedEntityList<>(
                 dtoList,
                 page,
-                (int) Math.ceil(departmentsCount / pageSize)
+                (int) Math.ceil((double) departmentsCount / pageSize)
         );
     }
