+++ app/src/main/java/t6proj/jobs/persistence/dao/JobDao.java
@@ -40,7 +40,7 @@ public class JobDao {
         return new PaginatedEntityList<>(
                 dtoList,
                 page,
-                (int) Math.ceil(departmentsCount / pageSize)
+                (int) Math.ceil((double) departmentsCount / pageSize)
         );
     }
